+++ lib/src/triggers/triggers.g.dart
@@ -70,6 +70,12 @@ _Trigger _$TriggerFromJson(Map<String, dynamic> json) => _Trigger(
       manualActionFields: (json['manualActionFields'] as List<dynamic>?)
           ?.map((e) => CareTask.fromJson(e as Map<String, dynamic>))
           .toList(),
+      children: (json['children'] as List<dynamic>?)
+          ?.map((e) => Trigger.fromJson(e as Map<String, dynamic>))
+          .toList(),
+      childrenIds: (json['childrenIds'] as List<dynamic>?)
+          ?.map((e) => e as String)
+          .toList(),
       access: (json['access'] as List<dynamic>?)
           ?.map((e) => Access.fromJson(e as Map<String, dynamic>))
           .toList(),
@@ -108,6 +114,8 @@ Map<String, dynamic> _$TriggerToJson(_Trigger instance) => <String, dynamic>{
           const DurationOrNullConverter().toJson(instance.cooldownTime),
       'manualActionFields':
           instance.manualActionFields?.map((e) => e.toJson()).toList(),
+      'children': instance.children?.map((e) => e.toJson()).toList(),
+      'childrenIds': instance.childrenIds,
       'access': instance.access?.map((e) => e.toJson()).toList(),
     };
 
@@ -121,6 +129,7 @@ const _$TriggerTypeEnumMap = {
   TriggerType.bhsSpeeding: 'BHS_SPEEDING',
   TriggerType.bhsPresence: 'BHS_PRESENCE',
   TriggerType.manualAction: 'MANUAL_ACTION',
+  TriggerType.nested: 'NESTED_TRIGGERS',
 };
 
 const _$TriggerGeofenceDetectionModeEnumMap = {
@@ -233,6 +242,10 @@ _TriggerInput _$TriggerInputFromJson(Map<String, dynamic> json) =>
               ?.map((e) => CareTaskInput.fromJson(e as Map<String, dynamic>))
               .toList() ??
           const [],
+      childrenIds: (json['childrenIds'] as List<dynamic>?)
+              ?.map((e) => e as String)
+              .toList() ??
+          const [],
       cooldownTime: json['cooldownTime'] == null
           ? const Duration(minutes: 1)
           : const DurationConverter().fromJson(json['cooldownTime'] as num),
@@ -269,6 +282,7 @@ Map<String, dynamic> _$TriggerInputToJson(_TriggerInput instance) =>
       'parameters': instance.parameters,
       'manualActionFields':
           instance.manualActionFields?.map((e) => e.toJson()).toList(),
+      'childrenIds': instance.childrenIds,
       'cooldownTime': const DurationConverter().toJson(instance.cooldownTime),
     };